minor fix and refactor
