+++ testFileProducer.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-sbt "project noah" "run-main edu.uci.ics.cloudberry.noah.feed.FileProducer \
+sbt "project noah" "run-main edu.uci.ics.cloudberry.noah.kafka.FileProducer \
 -kaf \
 -ks localhost:9092 \
 -fp ./noah/src/main/resources \
